Update Rest API in Preferences

+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/preferences/persistence/entity/ExciseAmphur.java
@@ -1,28 +1,22 @@
-
 package th.go.excise.ims.preferences.persistence.entity;
 
-import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.Table;
 import javax.persistence.Id;
+import javax.persistence.Table;
+
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
+
 import th.co.baiwa.buckwaframework.common.persistence.entity.BaseEntity;
 
 @Entity
 @Table(name = "EXCISE_AMPHUR")
 public class ExciseAmphur extends BaseEntity {
 
-
-
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = 6767121432761797009L;
 
 	public static class Field {
-
 		public static final String AMPHUR_ID = "AMPHUR_ID";
 		public static final String GEO_ID = "GEO_ID";
 		public static final String PROVINCE_ID = "PROVINCE_ID";
@@ -32,37 +26,37 @@ public class ExciseAmphur extends BaseEntity {
 
 	@Id
 	@Column(name = "AMPHUR_ID")
-	private BigDecimal amphurId;
+	private Long amphurId;
 	@Column(name = "GEO_ID")
-	private BigDecimal geoId;
+	private Long geoId;
 	@Column(name = "PROVINCE_ID")
-	private BigDecimal provinceId;
+	private Long provinceId;
 	@Column(name = "AMPHUR_CODE")
 	private String amphurCode;
 	@Column(name = "AMPHUR_NAME")
 	private String amphurName;
 
-	public BigDecimal getAmphurId() {
+	public Long getAmphurId() {
 		return amphurId;
 	}
 
-	public void setAmphurId(BigDecimal amphurId) {
+	public void setAmphurId(Long amphurId) {
 		this.amphurId = amphurId;
 	}
 
-	public BigDecimal getGeoId() {
+	public Long getGeoId() {
 		return geoId;
 	}
 
-	public void setGeoId(BigDecimal geoId) {
+	public void setGeoId(Long geoId) {
 		this.geoId = geoId;
 	}
 
-	public BigDecimal getProvinceId() {
+	public Long getProvinceId() {
 		return provinceId;
 	}
 
-	public void setProvinceId(BigDecimal provinceId) {
+	public void setProvinceId(Long provinceId) {
 		this.provinceId = provinceId;
 	}
